Fixing warnings

+++ source/common.m
@@ -46,7 +46,7 @@
             [result appendString:@">"];
         else if ([scanner scanString:@"&#" intoString:NULL]) {
             BOOL gotNumber;
-            unsigned charCode;
+            unsigned int charCode;
             NSString *xForHex = @"";
 			
             // Is it hex or decimal?
@@ -58,7 +58,7 @@
             }
 			
             if (gotNumber) {
-                [result appendFormat:@"%C", charCode];
+                [result appendFormat:@"%C", (unsigned short)charCode];
 				
 				[scanner scanString:@";" intoString:NULL];
             }
